+++ cmd/bundle/generate/job.go
@@ -31,38 +31,19 @@ func NewGenerateJobCommand() *cobra.Command {
 	cmd.Flags().Int64Var(&jobId, "existing-job-id", 0, `Job ID of the job to generate config for`)
 	cmd.MarkFlagRequired("existing-job-id")
 
-	cmd.Flags().StringVarP(&configDir, "config-dir", "d", "", `Dir path where the output config will be stored`)
-	cmd.Flags().StringVarP(&sourceDir, "source-dir", "s", "", `Dir path where the downloaded files will be stored`)
+	wd, err := os.Getwd()
+	if err != nil {
+		wd = "."
+	}
+
+	cmd.Flags().StringVarP(&configDir, "config-dir", "d", filepath.Join(wd, "resources"), `Dir path where the output config will be stored`)
+	cmd.Flags().StringVarP(&sourceDir, "source-dir", "s", filepath.Join(wd, "src"), `Dir path where the downloaded files will be stored`)
 	cmd.Flags().BoolVarP(&force, "force", "f", false, `Force overwrite existing files in the output directory`)
 
 	cmd.RunE = func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		b := bundle.Get(ctx)
 		w := b.WorkspaceClient()
-
-		if !cmd.Flag("config-dir").Changed {
-			wd, err := os.Getwd()
-			if err != nil {
-				return err
-			}
-			input, err := cmdio.Ask(ctx, "Output config dir", filepath.Join(wd, "resources"))
-			if err != nil {
-				return err
-			}
-			configDir = input
-		}
-
-		if !cmd.Flag("source-dir").Changed {
-			wd, err := os.Getwd()
-			if err != nil {
-				return err
-			}
-			input, err := cmdio.Ask(ctx, "Source code dir dir", filepath.Join(wd, "src"))
-			if err != nil {
-				return err
-			}
-			sourceDir = input
-		}
 
 		job, err := w.Jobs.Get(ctx, jobs.GetJobRequest{JobId: jobId})
 		if err != nil {
@@ -82,13 +63,13 @@ func NewGenerateJobCommand() *cobra.Command {
 			return err
 		}
 
-		jobKey := fmt.Sprintf("job_%s", textutil.NormaliseString(job.Settings.Name))
-		result := map[string]any{
-			"resources": map[string]any{
-				"jobs": map[string]dyn.Value{
+		jobKey := fmt.Sprintf("job_%s", textutil.NormalizeString(job.Settings.Name))
+		result := map[string]dyn.Value{
+			"resources": dyn.V(map[string]dyn.Value{
+				"jobs": dyn.V(map[string]dyn.Value{
 					jobKey: v,
-				},
-			},
+				}),
+			}),
 		}
 
 		err = downloader.FlushToDisk(ctx, force)
